# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: GmbH API
  version: 1.2.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/items:
    get:
      tags:
        - Item
      summary: Request all items
      responses:
        '200':
          description: Items list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemListResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Item
      summary: Create one item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreateRequest'
      responses:
        '200':
          description: Item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/items/{id}:
    get:
      tags:
        - Item
      summary: Request item by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Item
      summary: Update item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpdateRequest'
      responses:
        '200':
          description: Item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Item
      summary: Delete item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Empty response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/orderitems:
    get:
      tags:
        - Orderitem
      summary: Request all orderitems
      responses:
        '200':
          description: Orderitems list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderItem'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Orderitem
      summary: Create one orderitem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemCreateRequest'
      responses:
        '200':
          description: Orderitem
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderItemResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/orderitems/{id}:
    get:
      tags:
        - Orderitem
      summary: Request orderitem by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Orderitem
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderItemResponseLegacy'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Orderitem
      summary: Update orderitem
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemUpdateRequest'
      responses:
        '200':
          description: Orderitem
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderItemResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Orderitem
      summary: Delete orderitem
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Empty response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/orders:
    get:
      tags:
        - Order
      summary: Request all orders
      parameters:
        - in: query
          name: skip
          schema:
            type: integer
          required: false
        - in: query
          name: limit
          schema:
            type: integer
          required: false
      responses:
        '200':
          description: Orders list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Order
      summary: Create one order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/orders/byuser/{userId}:
    get:
      tags:
        - Order
      summary: Request all orders by user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: skip
          schema:
            type: integer
          required: false
        - in: query
          name: limit
          schema:
            type: integer
          required: false
      responses:
        '200':
          description: Orders list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/orders/{id}:
    get:
      tags:
        - Order
      summary: Request order by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Order
      summary: Update order by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Order
      summary: Delete order by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Empty response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-access-token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        errors:
          type: object
          properties:
            msg:
              type: string
          required:
            - msg
      required:
        - errors
...
    Item:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        amount:
          type: number
        price:
          type: number
        tax:
          type: number
        sort:
          type: number
        categoryId:
          type: string
        unitId:
          type: string
        enabled:
          type: boolean
    ItemResponse:
      type: object
      properties:
        item:
          $ref: '#/components/schemas/Item'
      required:
        - item
    ItemListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
      required:
        - items
    ItemCreateRequest:
      type: object
      properties:
        item:
          $ref: '#/components/schemas/Item'
      required:
        - item
    ItemUpdateRequest:
      type: object
      properties:
        item:
          $ref: '#/components/schemas/Item'
      required:
        - item
...
    OrderItem:
      type: object
      properties:
        id:
          type: string
        extras:
          type: string
        count:
          type: integer
        countPaid:
          type: integer
        countFree:
          type: integer
        price:
          type: number
        itemId:
          type: string
        orderId:
          type: string
    OrderItemResponse:
      type: object
      properties:
        orderitem:
          $ref: '#/components/schemas/OrderItem'
      required:
        - orderitem
    OrderItemResponseLegacy:
      type: object
      properties:
        item:
          $ref: '#/components/schemas/OrderItem'
      required:
        - item
    OrderItemCreateRequest:
      type: object
      properties:
        orderitem:
          $ref: '#/components/schemas/OrderItem'
      required:
        - orderitem
    OrderItemUpdateRequest:
      type: object
      properties:
        orderitem:
          $ref: '#/components/schemas/OrderItem'
      required:
        - orderitem
    Order:
      type: object
      properties:
        id:
          type: string
        totalAmount:
          type: number
        printCount:
          type: integer
        userId:
          type: string
        tableId:
          type: string
        orderitems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderResponse:
      type: object
      properties:
        order:
          $ref: '#/components/schemas/Order'
      required:
        - order
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        count:
          type: integer
        total:
          type: integer
      required:
        - orders
        - count
        - total
    OrderCreateRequest:
      type: object
      properties:
        order:
          $ref: '#/components/schemas/Order'
      required:
        - order
    OrderUpdateRequest:
      type: object
      properties:
        order:
          $ref: '#/components/schemas/Order'
      required:
        - order
...